Brought utd

+++ Nick/abl_mass_qsub_csstar.sh
@@ -7,6 +7,11 @@
 #$ -q short.q
 
 #Import the config file with shortcuts and settings
+# Import the config file with shortcuts and settings
+pwd
+if [[ ! -f "./config.sh" ]]; then
+	cp ./config_template.sh ./config.sh
+fi
 . ./config.sh
 #Import the module file that loads all necessary mods
 . "${mod_changers}/pipeline_mods"
@@ -21,7 +26,7 @@ if [[ $# -eq 0 ]]; then
 	exit 1
 # Shows a brief uasge/help section if -h option used as first argument
 elif [[ "$1" = "-h" ]]; then
-	echo "Usage is ./abl_mass_qsub_csstar.sh path_to_list_file(single sample ID per line, e.g. B8VHY/1700128 (it must include project id also)) max_concurrent_submissions"
+	echo "Usage is ./abl_mass_qsub_csstar.sh path_to_list_file(single sample ID per line, e.g. B8VHY/1700128 (it must include project id also)) max_submissions output_directory_for_scripts"
 	exit 1
 elif [[ ! -f "${1}" ]]; then
 	echo "${1} (list) does not exist...exiting"
@@ -43,21 +48,13 @@ echo "-${arr_size}:${arr[@]}-"
 counter=0
 max_subs=${2}
 
-# Set script directory
-main_dir="${share}/mass_subs/csstar_alt_subs"
-if [[ ! -d "${share}/mass_subs/csstar_alt_subs" ]]; then
-	mkdir "${share}/mass_subs/csstar_alt_subs"
-	mkdir "${share}/mass_subs/csstar_alt_subs/complete"
-elif [[ ! -d  "${share}/mass_subs/csstar_alt_subs/complete" ]]; then
-	mkdir "${share}/mass_subs/csstar_alt_subs/complete"
-fi
-
 # format name being extracted from alt database
-main_dir="${share}/mass_subs/csstar_subs"
-if [[ ! -d "${share}/mass_subs/csstar_subs" ]]; then
-	mkdir -p "${share}/mass_subs/csstar_subs/complete"
-elif [[ ! -d  "${share}/mass_subs/csstar_subs/complete" ]]; then
-	mkdir "${share}/mass_subs/csstar_subs/complete"
+main_dir="${3}/csstar_subs"
+cp ./config ${main_dir}
+if [[ ! -d "${3}/csstar_subs" ]]; then
+	mkdir -p "${3}/csstar_subs/complete"
+elif [[ ! -d  "${3}/csstar_subs/complete" ]]; then
+	mkdir "${3}/csstar_subs/complete"
 fi
 
 start_time=$(date "+%m-%d-%Y_at_%Hh_%Mm_%Ss")
@@ -66,13 +63,11 @@ start_time=$(date "+%m-%d-%Y_at_%Hh_%Mm_%Ss")
 while [ ${counter} -lt ${arr_size} ] ; do
 	sample=$(echo "${arr[${counter}]}" | cut -d'/' -f2)
 	project=$(echo "${arr[${counter}]}" | cut -d'/' -f1)
-	#rm -r ${processed}/${project}/${sample}/c-sstar/*20181003*
-	#rm -r ${processed}/${project}/${sample}/c-sstar_plasmid/*20181003*
 	echo ${counter}-${project}-${sample}
 	if [[ -s "${processed}/${project}/${sample}/Assembly/${sample}_scaffolds_trimmed.fasta" ]]; then
-		#echo "Test"
+		echo "Test"
 		if [[ ${counter} -lt ${max_subs} ]]; then
-			#if [[ ! -f "${processed}/${project}/${sample}/c-sstar/${sample}.${resGANNOT_srst2_filename}.gapped_98_sstar_summary.txt" ]]; then
+			if [[ ! -f "${processed}/${project}/${sample}/c-sstar/${sample}.${resGANNOT_srst2_filename}.gapped_98_sstar_summary.txt" ]]; then
 				echo  "Index is below max submissions, submitting"
 				echo -e "#!/bin/bash -l\n" > "${main_dir}/csstn_${sample}_${start_time}.sh"
 				echo -e "#$ -o csstn_${sample}.out" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
@@ -82,8 +77,11 @@ while [ ${counter} -lt ${arr_size} ] ; do
 				echo -e "#$ -q short.q\n"  >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 				echo -e "module load Python/3.6.1\n" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 				# Defaulting to gapped/98, change if you want to include user preferences
-				echo -e "\"${shareScript}/run_c-sstar_on_single.sh\" \"${sample}\" g l \"${project}\"" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
+				echo -e "cd ${shareScript}" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
+				echo -e "\"${shareScript}/run_c-sstar_on_single.sh\" \"${sample}\" g h \"${project}\"" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 				echo -e "echo \"$(date)\" > \"${main_dir}/complete/${sample}_csstarn_complete.txt\"" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
+				cd "${main_dir}"
+				echo "submitting ${main_dir}/csstn_${sample}_${start_time}.sh"
 				if [[ "${counter}" -lt "${last_index}" ]]; then
 					qsub "${main_dir}/csstn_${sample}_${start_time}.sh"
 				else
@@ -93,13 +91,14 @@ while [ ${counter} -lt ${arr_size} ] ; do
 						qsub -sync y "${main_dir}/csstn_${sample}_${start_time}.sh"
 					fi
 				fi
-			#else
-			#	echo "${project}/${sample} already has ${resGANNOT_srst2_filename}"
-			#	echo "$(date)" > "${main_dir}/complete/${sample}_csstarn_complete.txt"
-			#fi
-
+				mv "${shareScript}/csstn_${sample}.out" ${main_dir}
+				mv "${shareScript}/csstn_${sample}.err" ${main_dir}
+			else
+				echo "${project}/${sample} already has the newest ResGANNOT (${resGANNOT_srst2_filename})"
+				echo -e "$(date)" > "${main_dir}/complete/${sample}_csstarn_complete.txt"
+			fi
 			if [[ -d "${processed}/${project}/${sample}/c-sstar_plasmid" ]]; then
-				#if [[ ! -f "${processed}/${project}/${sample}/c-sstar_plasmid/${sample}.${resGANNOT_srst2_filename}.gapped_40_sstar_summary.txt" ]]; then
+				if [[ ! -f "${processed}/${project}/${sample}/c-sstar_plasmid/${sample}.${resGANNOT_srst2_filename}.gapped_40_sstar_summary.txt" ]]; then
 					echo "Index below max submissions, submitting plasmid"
 					echo -e "#!/bin/bash -l\n" > "${main_dir}/csstp_${sample}_${start_time}.sh"
 					echo -e "#$ -o csstp_${sample}.out" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
@@ -109,21 +108,22 @@ while [ ${counter} -lt ${arr_size} ] ; do
 					echo -e "#$ -q short.q\n"  >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 					echo -e "module load Python/3.6.1\n" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 					# Defaulting to gapped/98, change if you want to include user preferences
-					#echo -e "\"${shareScript}/fix_node_rename.sh\" \"${sample}\" \"${project}\"" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
+					echo -e "cd ${shareScript}" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 					echo -e "\"${shareScript}/run_c-sstar_on_single.sh\" \"${sample}\" g o \"${project}\" \"--plasmid\"" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 					echo -e "echo \"$(date)\" > \"${main_dir}/complete/${sample}_csstarp_complete.txt\"" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
+					cd "${main_dir}"
 					if [[ "${counter}" -lt "${last_index}" ]]; then
 						qsub "${main_dir}/csstp_${sample}_${start_time}.sh"
 					else
 						qsub -sync y "${main_dir}/csstp_${sample}_${start_time}.sh"
 					fi
-				#else
-				#	echo "${project}/${sample} already has ${resGANNOT_srst2_filename} PLASMID"
-				#	echo "$(date)" > "${main_dir}/complete/${sample}_csstarp_complete.txt"
-				#fi
+					mv "${shareScript}/csstp_${sample}.out" ${main_dir}
+					mv "${shareScript}/csstp_${sample}.err" ${main_dir}
+				else
+					echo "${project}/${sample} plasmid already has the newest ResGANNOT (${resGANNOT_srst2_filename})"
+					echo -e "$(date)" > "${main_dir}/complete/${sample}_csstarp_complete.txt"
+				fi
 			fi
-
-
 		else
 			waiting_for_index=$(( counter - max_subs ))
 			waiting_sample=$(echo "${arr[${waiting_for_index}]}" | cut -d'/' -f2)
@@ -146,8 +146,10 @@ while [ ${counter} -lt ${arr_size} ] ; do
 						echo -e "#$ -q short.q\n"  >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 						echo -e "module load Python/3.6.1\n" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 						# Defaulting to gapped/98, change if you want to include user preferences
+						echo -e "cd ${shareScript}" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 						echo -e "\"${shareScript}/run_c-sstar_on_single.sh\" \"${sample}\" g h \"${project}\"" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
 						echo -e "echo \"$(date)\" > \"${main_dir}/complete/${sample}_csstarn_complete.txt\"" >> "${main_dir}/csstn_${sample}_${start_time}.sh"
+						cd ${main_dir}
 						if [[ "${counter}" -lt "${last_index}" ]]; then
 							qsub "${main_dir}/csstn_${sample}_${start_time}.sh"
 						else
@@ -157,13 +159,14 @@ while [ ${counter} -lt ${arr_size} ] ; do
 								qsub -sync y "${main_dir}/csstn_${sample}_${start_time}.sh"
 							fi
 						fi
-					#else
-					#	echo "${project}/${sample} already has ${resGANNOT_srst2_filename}"
-					#	echo "$(date)" > "${main_dir}/complete/${sample}_csstarn_complete.txt"
-					#fi
-
+						mv "${shareScript}/csstn_${sample}.out" ${main_dir}
+						mv "${shareScript}/csstn_${sample}.err" ${main_dir}
+					else
+						echo "${project}/${sample} already has the newest ResGANNOT (${resGANNOT_srst2_filename})"
+						echo -e "$(date)" > "${main_dir}/complete/${sample}_csstarn_complete.txt"
+					fi
 					if [[ -d "${processed}/${project}/${sample}/c-sstar_plasmid" ]]; then
-						#if [[ ! -f "${processed}/${project}/${sample}/c-sstar_plasmid/${sample}.${resGANNOT_srst2_filename}.gapped_40_sstar_summary.txt" ]]; then
+						if [[ ! -f "${processed}/${project}/${sample}/c-sstar_plasmid/${sample}.${resGANNOT_srst2_filename}.gapped_40_sstar_summary.txt" ]]; then
 							echo -e "#!/bin/bash -l\n" > "${main_dir}/csstp_${sample}_${start_time}.sh"
 							echo -e "#$ -o csstp_${sample}.out" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 							echo -e "#$ -e csstp_${sample}.err" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
@@ -172,22 +175,26 @@ while [ ${counter} -lt ${arr_size} ] ; do
 							echo -e "#$ -q short.q\n"  >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 							echo -e "module load Python/3.6.1\n" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 							# Defaulting to gapped/98, change if you want to include user preferences
+							echo -e "cd ${shareScript}" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 							echo -e "\"${shareScript}/run_c-sstar_on_single.sh\" \"${sample}\" g o \"${project}\" \"--plasmid\"" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
 							echo -e "echo \"$(date)\" > \"${main_dir}/complete/${sample}_csstarp_complete.txt\"" >> "${main_dir}/csstp_${sample}_${start_time}.sh"
+							cd ${main_dir}
 							if [[ "${counter}" -lt "${last_index}" ]]; then
 								qsub "${main_dir}/csstp_${sample}_${start_time}.sh"
 							else
 								qsub -sync y "${main_dir}/csstp_${sample}_${start_time}.sh"
 							fi
-						#else
-						#	echo "${project}/${sample} already has ${resGANNOT_srst2_filename} PLASMID"
-						#	echo "$(date)" > "${main_dir}/complete/${sample}_csstarp_complete.txt"
-						#fi
+							mv "${shareScript}/csstp_${sample}.out" ${main_dir}
+							mv "${shareScript}/csstp_${sample}.err" ${main_dir}
+						else
+							echo "${project}/${sample} plasmid already has the newest ResGANNOT (${resGANNOT_srst2_filename})"
+							echo -e "$(date)" > "${main_dir}/complete/${sample}_csstarp_complete.txt"
+						fi
 					fi
 					break
 				else
 					timer=$(( timer + 5 ))
-					echo "sleeping for 5 seconds, so far slept for ${timer}"
+					echo "${main_dir}/complete/${waiting_sample}_csstarn_complete.txt not ready, sleeping for 5 seconds, so far slept for ${timer}"
 					sleep 5
 				fi
 			done
@@ -196,25 +203,30 @@ while [ ${counter} -lt ${arr_size} ] ; do
 	counter=$(( counter + 1 ))
 done
 
-# Check for completion of last sample
-finish_counter=0
-waiting_for_index=${last_index}
-waiting_sample=$(echo "${arr[${last_index}]}" | cut -d'/' -f2)
+# Check for completion of all samples
 timer=0
-while :
-do
-	if [[ ${timer} -gt 3600 ]]; then
-		echo "Timer exceeded limit of 3600 seconds = 60 minutes"
-		break
-	fi
+
+for item in "${arr[@]}"; do
+	waiting_sample=$(echo "${item}" | cut -d'/' -f2)
 	if [[ -f "${main_dir}/complete/${waiting_sample}_csstarn_complete.txt" ]] || [[ ! -s "${processed}/${project}/${waiting_sample}/Assembly/${waiting_sample}_scaffolds_trimmed.fasta" ]]; then
-		echo "All isolates completed"
-		printf "%s %s" "Act_by_list.sh has completed ${2}" "${global_end_time}" | mail -s "act_by_list complete" nvx4@cdc.gov
-		exit 0
+		echo "${item} is complete"
 	else
-		counter+$(( counter + 1))
-		timer=$(( timer + 5 ))
-		echo "sleeping for 5 seconds, so far slept for ${timer}"
-		sleep 5
+		while :
+		do
+				if [[ ${timer} -gt 3600 ]]; then
+					echo "Timer exceeded limit of 3600 seconds = 60 minutes"
+					exit 1
+				fi
+				if [[ -f "${main_dir}/complete/${waiting_sample}_csstarn_complete.txt" ]]; then
+					echo "${item} is complete"
+					break
+				else
+					timer=$(( timer + 5 ))
+					echo "sleeping for 5 seconds, so far slept for ${timer}"
+					sleep 5
+				fi
+		done
 	fi
 done
+
+exit 0
